fix: leverage typescript
